ubsan: fix signed left-shift for all 1 <<
only detected in aoso2, but still present in most files

+++ source/algos/bmh-sbndm.c
@@ -49,7 +49,7 @@ int search(unsigned char *x, int m, unsigned char *y, int n) {
   for (i = 0; i < SIGMA; i++)
     B[i] = 0;
   for (i = 0; i < m; i++)
-    B[x[m - i - 1]] |= (unsigned int)1 << (i + WORD - m);
+    B[x[m - i - 1]] |= 1U << (i + WORD - m);
   for (i = 0; i < m; i++)
     y[n + i] = x[i];
   D = B[x[m - 1]];
@@ -116,14 +116,14 @@ int search_large(unsigned char *x, int m, unsigned char *y, int n) {
   for (i = 0; i < SIGMA; i++)
     B[i] = 0;
   for (i = 0; i < m; i++)
-    B[x[m - i - 1]] |= (unsigned int)1 << (i + WORD - m);
+    B[x[m - i - 1]] |= 1U << (i + WORD - m);
   for (i = 0; i < m; i++)
     y[n + i] = x[i];
   D = B[x[m - 1]];
   j = 1;
   // shift = 1;
   for (i = m - 2; i > 0; i--, j++) {
-    // if (D & (1 << (m - 1)))
+    // if (D & (1U << (m - 1)))
     //  shift = j;
     D = (D << 1) & B[x[i]];
   }
