+++ source/algos/sbndm.c
@@ -45,14 +45,14 @@ int search(unsigned char *x, int m, unsigned char *y, int n) {
   for (i = 0; i < SIGMA; i++)
     B[i] = 0;
   for (i = 0; i < m; i++)
-    B[x[m - i - 1]] |= (unsigned int)1 << (i + WORD - m);
+    B[x[m - i - 1]] |= 1U << (i + WORD - m);
 
   last = m;
   s = (unsigned int)(~0) << (WORD - m);
   s = (unsigned int)(~0);
   for (i = m - 1; i >= 0; i--) {
     s &= B[x[i]];
-    if (s & ((unsigned int)1 << (WORD - 1))) {
+    if (s & (1U << (WORD - 1))) {
       if (i > 0)
         last = i;
     }
@@ -116,14 +116,14 @@ int search_large(unsigned char *x, int m, unsigned char *y, int n) {
   for (i = 0; i < SIGMA; i++)
     B[i] = 0;
   for (i = 0; i < m; i++)
-    B[x[m - i - 1]] |= (unsigned int)1 << (i + WORD - m);
+    B[x[m - i - 1]] |= 1U << (i + WORD - m);
 
   last = m;
   s = (unsigned int)(~0) << (WORD - m);
   s = (unsigned int)(~0);
   for (i = m - 1; i >= 0; i--) {
     s &= B[x[i]];
-    if (s & ((unsigned int)1 << (WORD - 1))) {
+    if (s & (1U << (WORD - 1))) {
       if (i > 0)
         last = i;
     }
